Rename "Series" to "Book"
Every collection has a plural collection.

For example:
``
Chapters -> [Chapter]
Groups -> [Group]
Volumes -> [Volume]
``

But there was none for `Series`. `Series` is singular and plural so only option was to name it something such as `SeriesCollection`. That would look out of place. By renaming it to Book, we can do `Books -> [Book]`