Add first API request: Get All Groups
The current request model is as such:

• A "Request" class acts as a host of the session task responsible for downloading the JSON payload.
• Specific API requests, such as Get All Groups, is then created as a convenience initializer of Request in an extension.
• The logic for each API is housed in its own file to make maintainability easier.

I do not have complete confidence in this model versus one that uses subclasses or protocols.

I chose this model because the API requests seemed simple enough to justify simple initializers. Once I get to the big show, such as series, then it may prove to be a poor design.