New design of request subcomponent
I am not intimately familiar with protocols in Swift coming from an Objective-C background so this is final(?) design for performing requests I have landed on. It uses class hierarchies (subclasses).

• Request is a generic base class for all requests. It hosts the completion block.
• RequestJSON is a specialized subclass of Request which is capable of dispatching session tasks and calling a function when those complete.

Specific requests, such as for all books, will then be a subclass of RequestJSON overriding the functions it needs to.