Further refine Request component
"Request" is now completely responsible for handling the NSURLSessionDataTask for the request. All logic related to this has been moved from "ReqeustJSON" to "Request".

The task created by "Request" calls to a taskCompleted() and taskFailed() function when it concludes. These functions can be overridden by subclasses to specialize the data.

The callback handler passed to "Request" is no longer visible outside itself.
Subclasses can call finalize() to conclude the request and inform callback handler. This makes it easier for the request to deconstruct itself as there is a concrete endpoint.

taskFailed() provides a default implementation which calls out to finalize() which means in most cases a subclass only needs to override taskCompleted()